+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,70 +1,5 @@
-import NextAuth, { NextAuthOptions } from 'next-auth';
-import GoogleProvider from 'next-auth/providers/google';
-import { GoogleAuthProvider, signInWithCredential } from 'firebase/auth';
-import { v4 as uuid } from 'uuid';
-import { auth } from '@/service/firebase/firebase-auth';
-import {
-  fetchUserFormEamil,
-  fetchUserFromUid,
-  sendUser,
-} from '@/service/firebase/firebase-firestore';
-import { User } from '@/types/user';
-import { createTime } from '@/utils/day';
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_ID ?? '',
-      clientSecret: process.env.GOOGLE_SECRET ?? '',
-      authorization: {
-        params: {
-          prompt: 'select_account',
-        },
-      },
-    }),
-  ],
-  callbacks: {
-    async signIn({ account }) {
-      const googleCredential = GoogleAuthProvider.credential(account?.id_token);
-      const userCredential = await signInWithCredential(auth, googleCredential);
-      const uid = userCredential.user.uid;
-      const email = userCredential.user.email ?? '';
-      const fetchedUser = await fetchUserFromUid(uid);
-
-      if (!fetchedUser) {
-        const user: User = {
-          id: uuid(),
-          uid,
-          email,
-          recommendPosts: [],
-          commentPosts: [],
-          createdAt: createTime(),
-        };
-
-        await sendUser(uid, user);
-      }
-
-      return true;
-    },
-    async session({ session }) {
-      const user = session.user;
-
-      const fetchedUser = await fetchUserFormEamil(user?.email ?? '');
-
-      if (fetchedUser) {
-        const { uid, email, name, profileImage } = fetchedUser[0];
-        session.user = {
-          uid,
-          email,
-          name,
-          profileImage,
-        };
-      }
-
-      return session;
-    },
-  },
-};
+import { authOptions } from '@/next-auth/options';
+import NextAuth from 'next-auth';
 
 const handler = NextAuth(authOptions);
